refactor: build config
