--- packages/bitcoin/index.ts
@@ -1,10 +1,0 @@
-export * from './src/bip322/bip322-utils';
-export * from './src/bip322/sign-message-bip322-bitcoinjs';
-export * from './src/bip322/sign-message-bip322';
-export * from './src/bitcoin-signer';
-export * from './src/bitcoin.network';
-export * from './src/bitcoin.utils';
-export * from './src/derivation-path.utils';
-export * from './src/p2tr-address-gen';
-export * from './src/p2wpkh-address-gen';
-export * from './src/p2wsh-p2sh-address-gen';
