+++ packages/bitcoin/package.json
@@ -1,30 +1,34 @@
 {
   "name": "@leather-wallet/bitcoin",
+  "author": "Leather.io contact@leather.io",
   "description": "Shared bitcoin utilities",
   "version": "0.6.6",
   "license": "MIT",
-  "type": "module",
-  "author": "Leather.io contact@leather.io",
-  "scripts": {
-    "format": "prettier . --write --ignore-path ../../.prettierignore",
-    "lint": "eslint . --fix --ignore-path ../../.eslintignore",
-    "lint:fix": "eslint . --fix --ignore-path ../../.eslintignore",
-    "test:coverage": "vitest run --coverage",
-    "test:unit": "vitest run",
-    "ts:build": "tsup index.ts --format esm && tsc --emitDeclarationOnly",
-    "ts:build:watch": "concurrently  \"tsc --build --watch --preserveWatchOutput --emitDeclarationOnly\"  \"tsup index.ts --format esm --watch \"  ",
-    "typecheck": "tsc --noEmit -p ./tsconfig.json"
-  },
   "homepage": "https://github.com/leather-wallet/mono/tree/dev/packages/bitcoin",
   "repository": {
     "type": "git",
     "url": "git@github.com:leather-wallet/mono.git",
     "directory": "packages/bitcoin"
   },
+  "type": "module",
+  "scripts": {
+    "format": "prettier . --write --ignore-path ../../.prettierignore",
+    "lint": "eslint . --fix --ignore-path ../../.eslintignore",
+    "lint:fix": "eslint . --fix --ignore-path ../../.eslintignore",
+    "test:coverage": "vitest run --coverage",
+    "test:unit": "vitest run",
+    "ts:build": "tsup",
+    "ts:build:watch": "tsup --watch",
+    "typecheck": "tsc --noEmit -p ./tsconfig.json"
+  },
+  "exports": {
+    ".": "./dist/index.js"
+  },
   "bugs": "https://github.com/leather-wallet/mono/issues",
   "dependencies": {
     "@bitcoinerlab/secp256k1": "1.0.2",
     "@leather-wallet/constants": "workspace:*",
+    "@leather-wallet/crypto": "workspace:*",
     "@leather-wallet/models": "workspace:*",
     "@leather-wallet/utils": "workspace:*",
     "@noble/hashes": "1.4.0",
@@ -46,20 +50,12 @@
     "@leather-wallet/rpc": "workspace:*",
     "@leather-wallet/tsconfig-config": "workspace:*",
     "@vitest/coverage-istanbul": "0.34.6",
-    "concurrently": "8.2.2",
     "eslint": "8.53.0",
     "prettier": "3.2.5",
     "tslib": "2.6.2",
-    "tsup": "8.0.1",
+    "tsup": "8.1.0",
     "typescript": "5.3.3",
     "vitest": "0.34.6"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.js"
-    }
   },
   "keywords": [
     "bitcoin",
