--- packages/bitcoin/src/derivation-path.utils.ts
@@ -1,8 +1,0 @@
-export enum DerivationPathDepth {
-  Root = 0,
-  Purpose = 1,
-  CoinType = 2,
-  Account = 3,
-  ChangeReceive = 4,
-  AddressIndex = 5,
-}
