+++ packages/bitcoin/src/index.ts
@@ -1,0 +1,9 @@
+export * from './bip322/bip322-utils';
+export * from './bip322/sign-message-bip322-bitcoinjs';
+export * from './bip322/sign-message-bip322';
+export * from './bitcoin-signer';
+export * from './bitcoin.network';
+export * from './bitcoin.utils';
+export * from './p2tr-address-gen';
+export * from './p2wpkh-address-gen';
+export * from './p2wsh-p2sh-address-gen';
