+++ packages/crypto/README.md
@@ -1,0 +1,3 @@
+# `@leather-wallet/crypto`
+
+This package contains all the generic crypto utils. Consumed by chain-specific packages.
