+++ packages/crypto/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "@leather-wallet/crypto",
+  "author": "leather-wallet",
+  "description": "Generic crypto utils package for Leather",
+  "version": "0.0.1",
+  "license": "MIT",
+  "scripts": {
+    "prepublish": "ts:build",
+    "test:unit": "vitest",
+    "ts:build": "tsup",
+    "ts:build:watch": "tsc --watch",
+    "typecheck": "tsc --noEmit"
+  },
+  "exports": {
+    ".": "./dist/index.mjs"
+  },
+  "dependencies": {
+    "@scure/bip32": "1.4.0",
+    "@scure/bip39": "1.3.0"
+  },
+  "devDependencies": {
+    "@vitest/coverage-istanbul": "0.34.6",
+    "tsup": "8.1.0",
+    "vitest": "0.34.6"
+  },
+  "files": [
+    "dist"
+  ],
+  "publishConfig": {
+    "access": "public"
+  }
+}
