+++ packages/crypto/src/derivation-path-utils.spec.ts
@@ -1,4 +1,4 @@
-import { extractAccountIndexFromPath, extractAddressIndexFromPath } from './bitcoin.utils';
+import { extractAccountIndexFromPath, extractAddressIndexFromPath } from './derivation-path-utils';
 
 describe(extractAddressIndexFromPath.name, () => {
   test('should extract the address index from a derivation path', () => {
