+++ packages/crypto/src/derivation-path-utils.ts
@@ -1,0 +1,25 @@
+export enum DerivationPathDepth {
+  Root = 0,
+  Purpose = 1,
+  CoinType = 2,
+  Account = 3,
+  ChangeReceive = 4,
+  AddressIndex = 5,
+}
+
+function extractSectionFromDerivationPath(depth: DerivationPathDepth) {
+  return (path: string) => {
+    const segments = path.split('/');
+    const accountNum = parseInt(segments[depth].replaceAll("'", ''), 10);
+    if (isNaN(accountNum)) throw new Error(`Cannot parse ${DerivationPathDepth[depth]} from path`);
+    return accountNum;
+  };
+}
+
+export const extractAccountIndexFromPath = extractSectionFromDerivationPath(
+  DerivationPathDepth.Account
+);
+
+export const extractAddressIndexFromPath = extractSectionFromDerivationPath(
+  DerivationPathDepth.AddressIndex
+);
