+++ packages/crypto/tsup.config.ts
@@ -1,0 +1,9 @@
+import { defineConfig } from 'tsup';
+
+export default defineConfig({
+  entry: ['src/index.ts'],
+  sourcemap: true,
+  clean: true,
+  dts: true,
+  format: 'esm',
+});
