+++ packages/crypto/vitest.config.js
@@ -1,5 +1,5 @@
 import { defineConfig } from 'vitest/config';
 
-import { defaultVitestUnitTestingConfig } from '../../config/vitest-configs';
+import { defaultVitestUnitTestingConfig } from '../../config/vitest-helpers';
 
 export default defineConfig({ ...defaultVitestUnitTestingConfig });
