--- packages/models/index.ts
@@ -1,15 +1,0 @@
-export * from './src/crypto-assets/crypto-asset-balance.model';
-export * from './src/crypto-assets/crypto-asset-info.model';
-export * from './src/crypto-assets/bitcoin/inscription.model';
-export * from './src/currencies.model';
-export * from './src/fees/bitcoin-fees.model';
-export * from './src/fees/fees.model';
-export * from './src/fees/stacks-fees.model';
-export * from './src/market.model';
-export * from './src/types';
-export * from './src/types.utils';
-export * from './src/money.model';
-export * from './src/network.model';
-export * from './src/transactions/bitcoin-transaction.model';
-export * from './src/transactions/stacks-transaction.model';
-export * from './src/utxo.model';
