+++ packages/query/package.json
@@ -1,10 +1,16 @@
 {
   "name": "@leather-wallet/query",
+  "author": "Leather.io contact@leather.io",
   "description": "Leather query",
   "version": "0.6.8",
   "license": "MIT",
+  "homepage": "https://github.com/leather-wallet/mono/tree/dev/packages/query",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:leather-wallet/mono.git",
+    "directory": "packages/query"
+  },
   "type": "module",
-  "author": "Leather.io contact@leather.io",
   "scripts": {
     "format": "prettier . --write --ignore-path ../../.prettierignore",
     "lint": "eslint . --ignore-path ../../.eslintignore",
@@ -15,11 +21,12 @@
     "ts:build:watch": "concurrently  \"tsc --build --watch --preserveWatchOutput --emitDeclarationOnly\"  \"tsup index.ts --format esm --watch \"  ",
     "typecheck": "tsc --noEmit -p ./tsconfig.json"
   },
-  "homepage": "https://github.com/leather-wallet/mono/tree/dev/packages/query",
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:leather-wallet/mono.git",
-    "directory": "packages/query"
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js",
+      "require": "./dist/index.js"
+    }
   },
   "bugs": "https://github.com/leather-wallet/mono/issues",
   "dependencies": {
@@ -60,16 +67,9 @@
     "eslint": "8.53.0",
     "jsdom": "22.1.0",
     "prettier": "3.2.5",
-    "tsup": "8.0.1",
+    "tsup": "8.1.0",
     "typescript": "5.3.3",
     "vitest": "0.34.6"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.js"
-    }
   },
   "keywords": [
     "leather",
