+++ packages/tokens/package.json
@@ -1,34 +1,32 @@
 {
   "name": "@leather-wallet/tokens",
+  "author": "Leather.io contact@leather.io",
   "description": "Shared design tokens",
   "version": "0.5.0",
   "license": "MIT",
-  "author": "Leather.io contact@leather.io",
-  "scripts": {
-    "lint": "eslint . --ignore-path ../../.eslintignore",
-    "prepublish": "pnpm ts:build",
-    "ts:build": "tsc --build",
-    "ts:build:watch": "tsc --build --watch --preserveWatchOutput",
-    "typecheck": "tsc --noEmit -p ./tsconfig.json"
-  },
   "homepage": "https://github.com/leather-wallet/mono/tree/dev/packages/tokens",
   "repository": {
     "type": "git",
     "url": "git@github.com:leather-wallet/mono.git",
     "directory": "packages/tokens"
   },
+  "scripts": {
+    "lint": "eslint . --ignore-path ../../.eslintignore",
+    "prepublish": "pnpm ts:build",
+    "ts:build": "tsup",
+    "ts:build:watch": "tsup --watch",
+    "typecheck": "tsc --noEmit"
+  },
+  "exports": {
+    ".": "./dist/index.mjs"
+  },
   "bugs": "https://github.com/leather-wallet/mono/issues",
   "devDependencies": {
     "@leather-wallet/eslint-config": "workspace:*",
     "@leather-wallet/prettier-config": "workspace:*",
     "@leather-wallet/tsconfig-config": "workspace:*",
-    "prettier": "3.2.5"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js"
-    }
+    "prettier": "3.2.5",
+    "tsup": "8.1.0"
   },
   "keywords": [
     "leather",
