+++ packages/tokens/tsconfig.json
@@ -4,6 +4,6 @@
     "rootDir": "src",
     "outDir": "dist"
   },
-  "include": ["src/**/*", ".*.ts"],
+  "include": ["src/**/*"],
   "exclude": []
 }
