+++ packages/ui/src/components/box/box.native.tsx
@@ -1,5 +1,5 @@
 import { createBox } from '@shopify/restyle';
 
-import { Theme } from '../../theme-native';
+import { type Theme } from '../../theme-native';
 
 export const Box = createBox<Theme>();
