+++ packages/ui/tsconfig.web.json
@@ -1,6 +1,8 @@
 {
   "extends": "../tsconfig-config/tsconfig.base",
   "compilerOptions": {
+    "composite": true,
+    "incremental": true,
     "jsx": "react-jsx",
     "lib": ["es2017", "DOM"],
     "baseUrl": ".",
