+++ packages/utils/package.json
@@ -1,25 +1,28 @@
 {
   "name": "@leather-wallet/utils",
+  "author": "Leather.io contact@leather.io",
   "description": "Shared bitcoin utilities",
   "version": "0.6.6",
   "license": "MIT",
+  "homepage": "https://github.com/leather-wallet/mono/tree/dev/packages/utils",
+  "repository": {
+    "type": "git",
+    "url": "git@github.com:leather-wallet/mono.git",
+    "directory": "packages/utils"
+  },
   "type": "module",
-  "author": "Leather.io contact@leather.io",
   "scripts": {
     "format": "prettier . --write --ignore-path ../../.prettierignore",
     "lint": "eslint . --ignore-path ../../.eslintignore",
     "lint:fix": "eslint . --fix --ignore-path ../../.eslintignore",
     "test:coverage": "vitest run --coverage",
     "test:unit": "vitest run",
-    "ts:build": "tsup index.ts --format esm && tsc --emitDeclarationOnly",
-    "ts:build:watch": "concurrently  \"tsc --build --watch --preserveWatchOutput --emitDeclarationOnly\"  \"tsup index.ts --format esm --watch \"  ",
-    "typecheck": "tsc --noEmit -p ./tsconfig.json"
+    "ts:build": "tsup",
+    "ts:build:watch": "tsup --watch",
+    "typecheck": "tsc --noEmit"
   },
-  "homepage": "https://github.com/leather-wallet/mono/tree/dev/packages/utils",
-  "repository": {
-    "type": "git",
-    "url": "git@github.com:leather-wallet/mono.git",
-    "directory": "packages/utils"
+  "exports": {
+    ".": "./dist/index.js"
   },
   "bugs": "https://github.com/leather-wallet/mono/issues",
   "dependencies": {
@@ -33,20 +36,11 @@
     "@leather-wallet/prettier-config": "workspace:*",
     "@leather-wallet/tsconfig-config": "workspace:*",
     "@vitest/coverage-istanbul": "0.34.6",
-    "concurrently": "8.2.2",
     "eslint": "8.53.0",
-    "jsdom": "22.1.0",
     "prettier": "3.2.5",
-    "tsup": "8.0.1",
+    "tsup": "8.1.0",
     "typescript": "5.3.3",
     "vitest": "0.34.6"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.js"
-    }
   },
   "keywords": [
     "leather",
