+++ packages/utils/src/money/format-money.ts
@@ -2,7 +2,7 @@ import { currencyDecimalsMap } from '@leather-wallet/constants';
 import type { Currencies, Money, NumType } from '@leather-wallet/models';
 import BigNumber from 'bignumber.js';
 
-import { isBigInt, isUndefined } from '../..';
+import { isBigInt, isUndefined } from '..';
 
 type KnownCurrencyDecimals = keyof typeof currencyDecimalsMap;
