+++ packages/utils/src/money/is-money.ts
@@ -1,7 +1,7 @@
 import { Money } from '@leather-wallet/models';
 import BigNumber from 'bignumber.js';
 
-import { isObject } from '../..';
+import { isObject } from '..';
 
 export function isMoney(val: unknown): val is Money {
   if (!isObject(val)) return false;
