+++ packages/utils/vitest.config.js
@@ -1,17 +1,5 @@
 import { defineConfig } from 'vitest/config';
 
-export default defineConfig({
-  test: {
-    include: 'src/**/*.spec.{ts,tsx}',
-    coverage: {
-      provider: 'istanbul',
-      reporter: ['text', 'json-summary', 'json', 'html'],
-      reportsDirectory: './coverage',
-    },
-    globals: true,
-    environment: 'node',
-    setupFiles: './unit-test.setup.js',
-    deps: { interopDefault: true },
-    silent: false,
-  },
-});
+import { defaultVitestUnitTestingConfig } from '../../config/vitest-configs';
+
+export default defineConfig({ ...defaultVitestUnitTestingConfig });
